Frontend Script Update

+++ roboshop/components/frontend.sh
@@ -23,7 +23,15 @@ set -e
 
 echo Frontend Setup
 yum install nginx -y
+if [ $? -ne 0 ]; then
+  echo "Nginx Install Failed"
+  exit
+fi
 curl -s -L -o /tmp/frontend.zip "https://github.com/roboshop-devops-project/frontend/archive/main.zip"
+if [ $? -ne 0 ]; then
+  echo "Nginx Install Failed"
+  exit
+fi
 cd /usr/share/nginx/html
 rm -rf *
 unzip /tmp/frontend.zip
